VSR PRNL Rework
In a similar line to the previous .prnls I did for stock and SXT.
The implementation is a lot less clean, however, due to the way the
mod's files are organized. There are probably stray files I've missed,
and there are .cfg files that haven't been removed due to containing
both parts to be pruned and parts not to be pruned.
I would not be surprised if there are things I have missed.